MIBE-59: refactor: user struggle MII files
