+++ frontend/src/panel/support_interface/script_support/components/LevelOfSupportDetails.tsx
@@ -1,12 +1,9 @@
-import {
-  levelsOfSupport,
-  LevelOfSupport,
-} from '../userStruggleSupport/userSupportMII';
 import { StateSetter } from '../../../models/utilTypes';
 import Stack from '@mui/material/Stack/Stack';
 import Typography from '@mui/material/Typography/Typography';
 import Slider from '@mui/material/Slider/Slider';
 import { LoSDescription } from '../userStruggleSupport/levelOfSupportUtils';
+import { LevelOfSupport, levelsOfSupport } from '../../../models/UserSupport';
 
 interface LevelOfSupportDetailsProps {
   levelOfSupport: LevelOfSupport;
